# feat: use bitwsise xor operations for getting orientation