### Changes
- Represent `Orientation` as enum with 4 primary variants plus an `Auto` variant
- The underlying type is a `u8` (one byte)
- Use the `u8` representation when not displaying the orientation because one byte is inexpensive
- Since a strand can be forward or reverse, we represent the boolean values as `1` and `0` (`u8`) so that we can later use them for faster bitwise operations
- The `HashMap`s with the keys of type `Orientation` are now represented as vectors with a maximum of 5 elements, each corresponding to a variant of `Orientation`