@rrwick 2024-02-07
Thanks for this! However, the formatting changes make it difficult for me to see exactly what has changed...

Have you profiled this code, before and after the change? If so, how much of a performance improvement did it make? If it's a significant improvement, then I'd be interested in merging this PR and would ask if you could remove the formatting changes. But if it's a small difference, then I'm okay with leaving orientations as strings, even if that's less efficient.