@lavafroth 2024-02-07
Sure, I will remove the formatting changes and share the benchmarks.