@lavafroth 2024-02-07
# Benchmarks

Benchmarks are performed using [hyperfine](https://github.com/sharkdp/hyperfine) on 1 million runs of respective test functions. This was done by compiling a binary in release mode that only ran the specific test function.

## `test_get_orientation`

Current implementation:

```
Time (mean ± σ):      8.212 s ±  0.075 s    [User: 8.180 s, System: 0.004 s]
Range (min … max):    8.108 s …  8.335 s    10 runs
```

PR implementation:

```
Time (mean ± σ):      7.892 s ±  0.116 s    [User: 7.863 s, System: 0.003 s]
Range (min … max):    7.811 s …  8.200 s    10 runs
```

## `test_auto_determine_orientation`

Current implementation:

```
Time (mean ± σ):      2.196 s ±  0.018 s    [User: 2.185 s, System: 0.001 s]
Range (min … max):    2.170 s …  2.226 s    10 runs
```

PR implementation:

```
Time (mean ± σ):     499.9 ms ±   9.1 ms    [User: 496.4 ms, System: 1.2 ms]
Range (min … max):   488.4 ms … 520.1 ms    10 runs
```