@rrwick 2024-02-08
Thank you for doing the formatting changes!

I tried this out and found a bug: the `auto_determine_orientation` function was returning `Orientation::Auto` when it should have been returning the best possible orientation, and this caused Polypolish to fail. I think I've fixed this, and I've also changed an inequality in `get_orientation` to make sure that the new code produces results identical to the old code.

I then ran some benchmarks on a full run of `polypolish filter`, but I saw essentially no improvement 😢 While your benchmarks show that the `auto_determine_orientation` is much faster, that function is only called once. Also, the bitwise operations, while efficient, make the code harder for me to understand.

Since this new implementation adds about 75 lines of code and doesn't seem to help performance, I'm going to opt against merging - sorry! I do appreciate your work on it, though.